+++ frame/ethereum/src/lib.rs
@@ -23,32 +23,37 @@
 // Ensure we're `no_std` when compiling for Wasm.
 #![cfg_attr(not(feature = "std"), no_std)]
 
+use codec::{Decode, Encode};
+use ethereum_types::{Bloom, BloomInput, H160, H256, H64, U256};
+use evm::ExitReason;
+use fp_consensus::{PostLog, PreLog, FRONTIER_ENGINE_ID};
+use fp_evm::CallOrCreateInfo;
+use fp_storage::PALLET_ETHEREUM_SCHEMA;
+use frame_support::ensure;
 use frame_support::{
-	decl_module, decl_storage, decl_error, decl_event,
-	traits::Get, traits::FindAuthor,
-	weights::{Pays, PostDispatchInfo, Weight},
+	decl_error, decl_event, decl_module, decl_storage,
 	dispatch::DispatchResultWithPostInfo,
+	traits::Get,
+	weights::{Pays, PostDispatchInfo, Weight},
+};
+use frame_system::ensure_none;
+use pallet_evm::{BlockHashMapping, FeeCalculator, GasWeightMapping, Runner};
+use sha3::{Digest, Keccak256};
+use sp_runtime::{
+	generic::DigestItem,
+	traits::{One, Saturating, UniqueSaturatedInto, Zero},
+	transaction_validity::{
+		InvalidTransaction, TransactionSource, TransactionValidity, ValidTransactionBuilder,
+	},
+	DispatchError,
 };
 use sp_std::prelude::*;
-use frame_system::ensure_none;
-use frame_support::ensure;
-use ethereum_types::{H160, H64, H256, U256, Bloom, BloomInput};
-use sp_runtime::{
-	transaction_validity::{
-		TransactionValidity, TransactionSource, InvalidTransaction, ValidTransactionBuilder,
-	},
-	generic::DigestItem, traits::UniqueSaturatedInto, DispatchError,
-};
-use evm::ExitReason;
-use fp_evm::CallOrCreateInfo;
-use pallet_evm::{Runner, GasWeightMapping, FeeCalculator};
-use sha3::{Digest, Keccak256};
-use codec::{Encode, Decode};
-use fp_consensus::{FRONTIER_ENGINE_ID, PostLog, PreLog};
-use fp_storage::PALLET_ETHEREUM_SCHEMA;
 
+pub use ethereum::{
+	BlockV0 as Block, LegacyTransactionMessage, Log, Receipt, TransactionAction,
+	TransactionV0 as Transaction,
+};
 pub use fp_rpc::TransactionStatus;
-pub use ethereum::{Transaction, Log, Block, Receipt, TransactionAction, TransactionMessage};
 
 #[cfg(all(feature = "std", test))]
 mod tests;
@@ -63,7 +68,7 @@ pub enum ReturnValue {
 }
 
 /// The schema version for Pallet Ethereum's storage
-#[derive(Clone, Debug, Encode, Decode, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Clone, Copy, Debug, Encode, Decode, PartialEq, Eq, PartialOrd, Ord)]
 pub enum EthereumStorageSchema {
 	Undefined,
 	V1,
@@ -88,11 +93,14 @@ impl Get<H256> for IntermediateStateRoot {
 }
 
 /// Configuration trait for Ethereum pallet.
-pub trait Config: frame_system::Config<Hash=H256> + pallet_balances::Config + pallet_timestamp::Config + pallet_evm::Config {
+pub trait Config:
+	frame_system::Config<Hash = H256>
+	+ pallet_balances::Config
+	+ pallet_timestamp::Config
+	+ pallet_evm::Config
+{
 	/// The overarching event type.
 	type Event: From<Event> + Into<<Self as frame_system::Config>::Event>;
-	/// Find author for Ethereum.
-	type FindAuthor: FindAuthor<H160>;
 	/// How Ethereum state root is calculated.
 	type StateRoot: Get<H256>;
 }
@@ -100,14 +108,16 @@ pub trait Config: frame_system::Config<Hash=H256> + pallet_balances::Config + pa
 decl_storage! {
 	trait Store for Module<T: Config> as Ethereum {
 		/// Current building block's transactions and receipts.
-		Pending: Vec<(ethereum::Transaction, TransactionStatus, ethereum::Receipt)>;
+		Pending: Vec<(Transaction, TransactionStatus, ethereum::Receipt)>;
 
 		/// The current Ethereum block.
-		CurrentBlock: Option<ethereum::Block>;
+		CurrentBlock: Option<ethereum::BlockV0>;
 		/// The current Ethereum receipts.
 		CurrentReceipts: Option<Vec<ethereum::Receipt>>;
 		/// The current transaction statuses.
 		CurrentTransactionStatuses: Option<Vec<TransactionStatus>>;
+		// Mapping for block number and hashes.
+		BlockHash: map hasher(twox_64_concat) U256 => H256;
 	}
 	add_extra_genesis {
 		build(|_config: &GenesisConfig| {
@@ -128,7 +138,6 @@ decl_event!(
 	}
 );
 
-
 decl_error! {
 	/// Ethereum pallet errors.
 	pub enum Error for Module<T: Config> {
@@ -147,7 +156,7 @@ decl_module! {
 
 		/// Transact an Ethereum transaction.
 		#[weight = <T as pallet_evm::Config>::GasWeightMapping::gas_to_weight(transaction.gas_limit.unique_saturated_into())]
-		fn transact(origin, transaction: ethereum::Transaction) -> DispatchResultWithPostInfo {
+		fn transact(origin, transaction: Transaction) -> DispatchResultWithPostInfo {
 			ensure_none(origin)?;
 
 			Self::do_transact(transaction)
@@ -162,9 +171,18 @@ decl_module! {
 					)
 				),
 			);
+			// move block hash pruning window by one block
+			let block_hash_count = T::BlockHashCount::get();
+			let to_remove = n.saturating_sub(block_hash_count).saturating_sub(One::one());
+			// keep genesis hash
+			if !to_remove.is_zero() {
+				BlockHash::remove(U256::from(
+					UniqueSaturatedInto::<u32>::unique_saturated_into(to_remove)
+				));
+			}
 		}
 
-		fn on_initialize(n: T::BlockNumber) -> Weight {
+		fn on_initialize(_n: T::BlockNumber) -> Weight {
 			Pending::kill();
 
 			if let Ok(log) = fp_consensus::find_pre_log(&frame_system::Module::<T>::digest()) {
@@ -178,6 +196,14 @@ decl_module! {
 
 			0
 		}
+	}
+}
+
+/// Returns the Ethereum block hash by number.
+pub struct EthereumBlockHashMapping;
+impl BlockHashMapping for EthereumBlockHashMapping {
+	fn block_hash(number: u32) -> H256 {
+		BlockHash::get(U256::from(number))
 	}
 }
 
@@ -197,15 +223,22 @@ impl<T: Config> frame_support::unsigned::ValidateUnsigned for Module<T> {
 		if let Call::transact(transaction) = call {
 			if let Some(chain_id) = transaction.signature.chain_id() {
 				if chain_id != T::ChainId::get() {
-					return InvalidTransaction::Custom(TransactionValidationError::InvalidChainId as u8).into();
+					return InvalidTransaction::Custom(
+						TransactionValidationError::InvalidChainId as u8,
+					)
+					.into();
 				}
 			}
 
-			let origin = Self::recover_signer(&transaction)
-				.ok_or_else(|| InvalidTransaction::Custom(TransactionValidationError::InvalidSignature as u8))?;
+			let origin = Self::recover_signer(&transaction).ok_or_else(|| {
+				InvalidTransaction::Custom(TransactionValidationError::InvalidSignature as u8)
+			})?;
 
 			if transaction.gas_limit >= T::BlockGasLimit::get() {
-				return InvalidTransaction::Custom(TransactionValidationError::InvalidGasLimit as u8).into();
+				return InvalidTransaction::Custom(
+					TransactionValidationError::InvalidGasLimit as u8,
+				)
+				.into();
 			}
 
 			let account_data = pallet_evm::Module::<T>::account_basic(&origin);
@@ -228,12 +261,7 @@ impl<T: Config> frame_support::unsigned::ValidateUnsigned for Module<T> {
 
 			let mut builder = ValidTransactionBuilder::default()
 				.and_provides((origin, transaction.nonce))
-				.priority(if min_gas_price == U256::zero() {
-						0
-					} else {
-						let target_gas = (transaction.gas_limit * transaction.gas_price) / min_gas_price;
-						T::GasWeightMapping::gas_to_weight(target_gas.unique_saturated_into())
-				});
+				.priority(transaction.gas_price.unique_saturated_into());
 
 			if transaction.nonce > account_data.nonce {
 				if let Some(prev_nonce) = transaction.nonce.checked_sub(1.into()) {
@@ -249,16 +277,18 @@ impl<T: Config> frame_support::unsigned::ValidateUnsigned for Module<T> {
 }
 
 impl<T: Config> Module<T> {
-	fn recover_signer(transaction: &ethereum::Transaction) -> Option<H160> {
+	fn recover_signer(transaction: &Transaction) -> Option<H160> {
 		let mut sig = [0u8; 65];
 		let mut msg = [0u8; 32];
 		sig[0..32].copy_from_slice(&transaction.signature.r()[..]);
 		sig[32..64].copy_from_slice(&transaction.signature.s()[..]);
 		sig[64] = transaction.signature.standard_v();
-		msg.copy_from_slice(&TransactionMessage::from(transaction.clone()).hash()[..]);
+		msg.copy_from_slice(&LegacyTransactionMessage::from(transaction.clone()).hash()[..]);
 
 		let pubkey = sp_io::crypto::secp256k1_ecdsa_recover(&sig, &msg).ok()?;
-		Some(H160::from(H256::from_slice(Keccak256::digest(&pubkey).as_slice())))
+		Some(H160::from(H256::from_slice(
+			Keccak256::digest(&pubkey).as_slice(),
+		)))
 	}
 
 	fn store_block(post_log: bool, block_number: U256) {
@@ -270,39 +300,38 @@ impl<T: Config> Module<T> {
 			transactions.push(transaction);
 			statuses.push(status);
 			receipts.push(receipt.clone());
-			Self::logs_bloom(
-				receipt.logs.clone(),
-				&mut logs_bloom
-			);
+			Self::logs_bloom(receipt.logs.clone(), &mut logs_bloom);
 		}
 
 		let ommers = Vec::<ethereum::Header>::new();
+		let receipts_root =
+			ethereum::util::ordered_trie_root(receipts.iter().map(|r| rlp::encode(r)));
 		let partial_header = ethereum::PartialHeader {
 			parent_hash: Self::current_block_hash().unwrap_or_default(),
-			beneficiary: <Module<T>>::find_author(),
-			// TODO: figure out if there's better way to get a sort-of-valid state root.
-			state_root: H256::default(),
-			receipts_root: H256::from_slice(
-				Keccak256::digest(&rlp::encode_list(&receipts)[..]).as_slice(),
-			), // TODO: check receipts hash.
+			beneficiary: pallet_evm::Module::<T>::find_author(),
+			state_root: T::StateRoot::get(),
+			receipts_root,
 			logs_bloom,
 			difficulty: U256::zero(),
 			number: block_number,
 			gas_limit: T::BlockGasLimit::get(),
-			gas_used: receipts.clone().into_iter().fold(U256::zero(), |acc, r| acc + r.used_gas),
+			gas_used: receipts
+				.clone()
+				.into_iter()
+				.fold(U256::zero(), |acc, r| acc + r.used_gas),
 			timestamp: UniqueSaturatedInto::<u64>::unique_saturated_into(
-				pallet_timestamp::Module::<T>::get()
+				pallet_timestamp::Module::<T>::get(),
 			),
 			extra_data: Vec::new(),
 			mix_hash: H256::default(),
 			nonce: H64::default(),
 		};
-		let mut block = ethereum::Block::new(partial_header, transactions.clone(), ommers);
-		block.header.state_root = T::StateRoot::get();
+		let block = ethereum::Block::new(partial_header, transactions.clone(), ommers);
 
 		CurrentBlock::put(block.clone());
 		CurrentReceipts::put(receipts.clone());
 		CurrentTransactionStatuses::put(statuses.clone());
+		BlockHash::insert(block_number, block.header.hash());
 
 		if post_log {
 			let digest = DigestItem::<T::Hash>::Consensus(
@@ -322,18 +351,17 @@ impl<T: Config> Module<T> {
 		}
 	}
 
-	fn do_transact(transaction: ethereum::Transaction) -> DispatchResultWithPostInfo {
+	fn do_transact(transaction: Transaction) -> DispatchResultWithPostInfo {
 		ensure!(
 			fp_consensus::find_pre_log(&frame_system::Module::<T>::digest()).is_err(),
 			Error::<T>::PreLogExists,
 		);
 
-		let source = Self::recover_signer(&transaction)
-			.ok_or_else(|| Error::<T>::InvalidSignature)?;
+		let source =
+			Self::recover_signer(&transaction).ok_or_else(|| Error::<T>::InvalidSignature)?;
 
-		let transaction_hash = H256::from_slice(
-			Keccak256::digest(&rlp::encode(&transaction)).as_slice()
-		);
+		let transaction_hash =
+			H256::from_slice(Keccak256::digest(&rlp::encode(&transaction)).as_slice());
 		let transaction_index = Pending::get().len() as u32;
 
 		let (to, contract_address, info) = Self::execute(
@@ -348,8 +376,9 @@ impl<T: Config> Module<T> {
 		)?;
 
 		let (reason, status, used_gas) = match info {
-			CallOrCreateInfo::Call(info) => {
-				(info.exit_reason, TransactionStatus {
+			CallOrCreateInfo::Call(info) => (
+				info.exit_reason,
+				TransactionStatus {
 					transaction_hash,
 					transaction_index,
 					from: source,
@@ -358,16 +387,15 @@ impl<T: Config> Module<T> {
 					logs: info.logs.clone(),
 					logs_bloom: {
 						let mut bloom: Bloom = Bloom::default();
-						Self::logs_bloom(
-							info.logs,
-							&mut bloom
-						);
+						Self::logs_bloom(info.logs, &mut bloom);
 						bloom
 					},
-				}, info.used_gas)
-			},
-			CallOrCreateInfo::Create(info) => {
-				(info.exit_reason, TransactionStatus {
+				},
+				info.used_gas,
+			),
+			CallOrCreateInfo::Create(info) => (
+				info.exit_reason,
+				TransactionStatus {
 					transaction_hash,
 					transaction_index,
 					from: source,
@@ -376,14 +404,12 @@ impl<T: Config> Module<T> {
 					logs: info.logs.clone(),
 					logs_bloom: {
 						let mut bloom: Bloom = Bloom::default();
-						Self::logs_bloom(
-							info.logs,
-							&mut bloom
-						);
+						Self::logs_bloom(info.logs, &mut bloom);
 						bloom
 					},
-				}, info.used_gas)
-			},
+				},
+				info.used_gas,
+			),
 		};
 
 		let receipt = ethereum::Receipt {
@@ -400,19 +426,19 @@ impl<T: Config> Module<T> {
 
 		Pending::append((transaction, status, receipt));
 
-		Self::deposit_event(Event::Executed(source, contract_address.unwrap_or_default(), transaction_hash, reason));
+		Self::deposit_event(Event::Executed(
+			source,
+			contract_address.unwrap_or_default(),
+			transaction_hash,
+			reason,
+		));
 		Ok(PostDispatchInfo {
-			actual_weight: Some(T::GasWeightMapping::gas_to_weight(used_gas.unique_saturated_into())),
+			actual_weight: Some(T::GasWeightMapping::gas_to_weight(
+				used_gas.unique_saturated_into(),
+			)),
 			pays_fee: Pays::No,
-		}).into()
-	}
-
-	/// Get the author using the FindAuthor trait.
-	pub fn find_author() -> H160 {
-		let digest = <frame_system::Module<T>>::digest();
-		let pre_runtime_digests = digest.logs.iter().filter_map(|d| d.as_pre_runtime());
-
-		T::FindAuthor::find_author(pre_runtime_digests).unwrap_or_default()
+		})
+		.into()
 	}
 
 	/// Get the transaction status with given index.
@@ -421,7 +447,7 @@ impl<T: Config> Module<T> {
 	}
 
 	/// Get current block.
-	pub fn current_block() -> Option<ethereum::Block> {
+	pub fn current_block() -> Option<ethereum::BlockV0> {
 		CurrentBlock::get()
 	}
 
@@ -457,10 +483,11 @@ impl<T: Config> Module<T> {
 					gas_price,
 					nonce,
 					config.as_ref().unwrap_or(T::config()),
-				).map_err(Into::into)?;
+				)
+				.map_err(Into::into)?;
 
 				Ok((Some(target), None, CallOrCreateInfo::Call(res)))
-			},
+			}
 			ethereum::TransactionAction::Create => {
 				let res = T::Runner::create(
 					from,
@@ -470,10 +497,11 @@ impl<T: Config> Module<T> {
 					gas_price,
 					nonce,
 					config.as_ref().unwrap_or(T::config()),
-				).map_err(Into::into)?;
+				)
+				.map_err(Into::into)?;
 
 				Ok((None, Some(res.value), CallOrCreateInfo::Create(res)))
-			},
+			}
 		}
 	}
 }