+++ frame/evm/test-vector-support/src/lib.rs
@@ -15,8 +15,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use evm::{executor::PrecompileOutput, Context, ExitSucceed};
 use fp_evm::Precompile;
-use evm::{ExitSucceed, Context, executor::PrecompileOutput};
 
 #[cfg(feature = "std")]
 use serde::Deserialize;
@@ -35,20 +35,17 @@ struct EthConsensusTest {
 /// The file is expected to be in JSON format and contain an array of test vectors, where each
 /// vector can be deserialized into an "EthConsensusTest".
 #[cfg(feature = "std")]
-pub fn test_precompile_test_vectors<P: Precompile>(filepath: &str)
-	-> std::result::Result<(), String>
-{
+pub fn test_precompile_test_vectors<P: Precompile>(
+	filepath: &str,
+) -> std::result::Result<(), String> {
 	use std::fs;
 
-	let data = fs::read_to_string(&filepath)
-		.expect("Failed to read blake2F.json");
+	let data = fs::read_to_string(&filepath).expect("Failed to read blake2F.json");
 
-	let tests: Vec<EthConsensusTest> = serde_json::from_str(&data)
-		.expect("expected json array");
+	let tests: Vec<EthConsensusTest> = serde_json::from_str(&data).expect("expected json array");
 
 	for test in tests {
-		let input: Vec<u8> = hex::decode(test.Input)
-			.expect("Could not hex-decode test input data");
+		let input: Vec<u8> = hex::decode(test.Input).expect("Could not hex-decode test input data");
 
 		let cost: u64 = 10000000;
 
@@ -61,15 +58,26 @@ pub fn test_precompile_test_vectors<P: Precompile>(filepath: &str)
 		match P::execute(&input, Some(cost), &context) {
 			Ok(result) => {
 				let as_hex: String = hex::encode(result.output);
-				assert_eq!(result.exit_status, ExitSucceed::Returned,
-						"test '{}' returned {:?} (expected 'Returned')", test.Name, result.exit_status);
-				assert_eq!(as_hex, test.Expected,
-						"test '{}' failed (different output)", test.Name);
+				assert_eq!(
+					result.exit_status,
+					ExitSucceed::Returned,
+					"test '{}' returned {:?} (expected 'Returned')",
+					test.Name,
+					result.exit_status
+				);
+				assert_eq!(
+					as_hex, test.Expected,
+					"test '{}' failed (different output)",
+					test.Name
+				);
 				if let Some(expected_gas) = test.Gas {
-					assert_eq!(result.cost, expected_gas,
-							"test '{}' failed (different gas cost)", test.Name);
+					assert_eq!(
+						result.cost, expected_gas,
+						"test '{}' failed (different gas cost)",
+						test.Name
+					);
 				}
-			},
+			}
 			Err(err) => {
 				return Err(format!("Test '{}' returned error: {:?}", test.Name, err));
 			}