+++ client/db/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog for `fc-db`
 
 ## Unreleased
+* Introduce versioning and migration functions.
+* New version 2.
+* New column `ETHEREUM_SCHEMA_CACHE` in version 2.
+* New dependency  `fp-storage`.
+* New dependency `pallet-ethereum`.